requirements.txt
ollama
requests
pyttsx3
SpeechRecognition
groq
customtkinter
Pillow
pygame
pyautogui
screen-brightness-control
